refactor: store_modules directory struct
