+++ src/store_modules/todo/components/list.vue
@@ -47,12 +47,12 @@
 
 <script lang='ts'>
 import { Component, Vue, Prop, Watch } from 'nuxt-property-decorator';
-import { TodoState } from '../store/state';
-import Todo from '../store/state/todo';
-import { TodoActions } from '../store/actions';
+import { TodoState } from '../state';
+import Todo from '../state/todo';
+import { TodoActions } from '../actions';
 import { HTMLElementEvent } from '../../../types';
 import { Actions, Getters } from '../../../mixins/store_helper';
-import { TodoGetters } from '../store/getters';
+import { TodoGetters } from '../getters';
 import Item from './item.vue';
 import VueI18n from 'vue-i18n';
 import { Route } from 'vue-router';
